Rename osm object base class
