+++ ts/src/osm-object.ts
